Validate posts and create form partial.
Validate posts to require a minimum title and body length and require
an associated topic and user. Create a partial called form that has
title and body text fields and. Render them on posts edit and new views
while passing topic and post instances to local variables. Define the
form_group_tag helper which adds the 'has-error' class to divs if posts
are not valid. Update form partial to display list of errors if post entry
is not valid.